Frontend: Added the paginator!!!
Ah, Perry the Platypus, you've arrived just in time to witness my latest and greatest evil invention: the Paginator! You see, in the past, my -inators have had some rather, let's say, unique and peculiar purposes. But today, I present to you something that will truly baffle and bewilder the world!

The Paginator, my dear nemesis, is a device that will allow me to control time itself. Yes, you heard me correctly - time manipulation! With this magnificent contraption, I will have the power to accelerate and decelerate the passage of time, wreaking havoc on the very fabric of reality.

Imagine, Perry, the chaos I could unleash! I could speed up time, making people age in the blink of an eye, or slow it down to a crawl, trapping them in perpetual boredom. I could age cheese to perfection, or perhaps accelerate the growth of inedible fruit. The possibilities are as endless as my thirst for evil!

But that's not all, my platypus adversary. The Paginator can also rewind time, allowing me to correct past mistakes, or more importantly, undo your thwarting of my -inators. No longer will you be able to stop me in my tracks, for I will simply rewind and try again. It's the ultimate do-over!

Of course, every great invention needs a self-destruct button, but not this one! I've learned from my past mistakes, and the Paginator is completely foolproof. It's a flawless creation, and I don't foresee any unintended consequences. Well, perhaps a few, but I'll deal with those later.

So, Perry the Platypus, prepare to witness the dawn of a new era of evil! With the Paginator in my grasp, I'll be unstoppable. And soon, all of Danville will be under my control! Mwahahaha!