Frontend: Color armageddon
This is a HUGE commit, and it needed to be big to get the basic infrastructure in place. Basically, our website now supports 2 different styles with opposing color schemes via css with some tinkering. Anywhere in the code it says "fight" can be replaced via jsx with either depending on the state. But the idea is, most of the heavy lifting of color schemes and whatnot can be taken care of through scss rather than individual css styles, and any extra stuff can easily reference this scss. More work will have to be done with this obviously, but it should put us in a good position to more easily tinker with it on a larger scale. (Most of this is directly modified and was taken right from the dark color mode implementation, since that was recommend in documentation as a template for other color modes.)